+++ src/datasources/asset/components/variable-editor/AssetVariableQueryEditor.test.tsx
@@ -1,11 +1,12 @@
-import { screen, waitFor } from '@testing-library/react';
-import { AssetQuery, AssetQueryType, QueryReturnType } from '../../types/types';
+import { act, screen, waitFor } from '@testing-library/react';
+import { AssetQueryType, QueryReturnType } from '../../types/types';
 import { SystemProperties } from '../../../system/types'
 import { AssetVariableQueryEditor } from './AssetVariableQueryEditor';
 import { Workspace } from 'core/types';
 import { setupRenderer } from 'test/fixtures';
 import { ListAssetsDataSource } from '../../data-sources/list-assets/ListAssetsDataSource';
 import { AssetDataSource } from 'datasources/asset/AssetDataSource';
+import userEvent from '@testing-library/user-event';
 import { select } from 'react-select-event';
 import { AssetVariableQuery } from 'datasources/asset/types/AssetVariableQuery.types';
 
@@ -52,15 +53,170 @@ class FakeAssetDataSource extends AssetDataSource {
     }
 }
 
-const render = setupRenderer(AssetVariableQueryEditor, FakeAssetDataSource, () => {});
+const render = async (query: AssetVariableQuery) => {
+    return await act(async () => setupRenderer(AssetVariableQueryEditor, FakeAssetDataSource, () => { })(query))
+}
+
 
 it('renders the variable query builder', async () => {
-    render({  refId: '', type: AssetQueryType.ListAssets, filter: "" } as AssetQuery);
+    render({ refId: '', type: AssetQueryType.ListAssets, filter: "" } as AssetVariableQuery);
 
     await waitFor(() => expect(screen.getAllByText('Property').length).toBe(1));
     await waitFor(() => expect(screen.getAllByText('Operator').length).toBe(1));
     await waitFor(() => expect(screen.getAllByText('Value').length).toBe(1));
 });
+
+it('should render take', async () => {
+    await render({ type: AssetQueryType.ListAssets } as AssetVariableQuery)
+    expect(screen.queryByRole('spinbutton')).toBeInTheDocument();
+});
+
+it('only allows numbers input in Take field', async () => {
+    await render({ type: AssetQueryType.ListAssets, take: 1000 } as AssetVariableQuery);
+
+    const take = screen.getByRole('spinbutton');
+
+    await userEvent.clear(take);
+    await userEvent.type(take, 'aaa');
+    await waitFor(() => {
+        expect(take).toHaveValue(null);
+    });
+
+    await userEvent.clear(take);
+    await userEvent.type(take, '5');
+    await waitFor(() => {
+        expect(take).toHaveValue(5);
+    });
+})
+
+it('should call onChange with take when user changes take', async () => {
+    const [onChange] = await render({ type: AssetQueryType.ListAssets, take: 1000 } as AssetVariableQuery);
+    const take = screen.getByRole('spinbutton');
+
+    await userEvent.clear(take);
+    await userEvent.type(take, '5');
+    await userEvent.tab();
+    await waitFor(() => {
+        expect(onChange).toHaveBeenCalledWith(expect.objectContaining({ take: 5 }));
+    });
+})
+
+it('should display error message when user changes take to number greater than max take', async () => {
+    const [onChange] = await render({} as AssetVariableQuery);
+    const take = screen.getByRole('spinbutton');
+    onChange.mockClear();
+
+    await userEvent.clear(take);
+    await userEvent.type(take, '10001');
+    await userEvent.tab();
+    await waitFor(() => {
+        expect(screen.getByText('Enter a value less than or equal to 10,000')).toBeInTheDocument();
+        expect(onChange).toHaveBeenCalledWith(expect.objectContaining({ take: undefined }));
+    });
+})
+
+it('should display error message when user changes take to number less than 0', async () => {
+    const [onChange] = await render({} as AssetVariableQuery);
+    const take = screen.getByRole('spinbutton');
+    onChange.mockClear();
+
+    await userEvent.clear(take);
+    await userEvent.tab();
+    await waitFor(() => {
+        expect(screen.getByText('Enter a value greater than or equal to 0')).toBeInTheDocument();
+        expect(onChange).toHaveBeenCalledWith(expect.objectContaining({ take: undefined }));
+    });
+})
+
+it('should not display error message when user changes value to number between 0 and max take', async () => {
+    const [onChange] = await render({} as AssetVariableQuery);
+    const take = screen.getByRole('spinbutton');
+    onChange.mockClear();
+
+    await userEvent.clear(take);
+    await userEvent.type(take, '5')
+    await userEvent.tab();
+    await waitFor(() => {
+        expect(screen.queryByText('Enter a value greater than or equal to 0')).not.toBeInTheDocument();
+        expect(screen.queryByText('Enter a value less than or equal to 10,000')).not.toBeInTheDocument();
+    });
+})
+
+it('should render take', async () => {
+    await render({ type: AssetQueryType.ListAssets } as AssetVariableQuery)
+    expect(screen.queryByRole('spinbutton')).toBeInTheDocument();
+});
+
+it('only allows numbers input in Take field', async () => {
+    await render({ type: AssetQueryType.ListAssets, take: 1000 } as AssetVariableQuery);
+
+    const take = screen.getByRole('spinbutton');
+
+    await userEvent.clear(take);
+    await userEvent.type(take, 'aaa');
+    await waitFor(() => {
+        expect(take).toHaveValue(null);
+    });
+
+    await userEvent.clear(take);
+    await userEvent.type(take, '5');
+    await waitFor(() => {
+        expect(take).toHaveValue(5);
+    });
+})
+
+it('should call onChange with take when user changes take', async () => {
+    const [onChange] = await render({ type: AssetQueryType.ListAssets, take: 1000 } as AssetVariableQuery);
+    const take = screen.getByRole('spinbutton');
+
+    await userEvent.clear(take);
+    await userEvent.type(take, '5');
+    await userEvent.tab();
+    await waitFor(() => {
+        expect(onChange).toHaveBeenCalledWith(expect.objectContaining({ take: 5 }));
+    });
+})
+
+it('should display error message when user changes take to number greater than max take', async () => {
+    const [onChange] = await render({} as AssetVariableQuery);
+    const take = screen.getByRole('spinbutton');
+    onChange.mockClear();
+
+    await userEvent.clear(take);
+    await userEvent.type(take, '10001');
+    await userEvent.tab();
+    await waitFor(() => {
+        expect(screen.getByText('Enter a value less than or equal to 10,000')).toBeInTheDocument();
+        expect(onChange).toHaveBeenCalledWith(expect.objectContaining({ take: undefined }));
+    });
+})
+
+it('should display error message when user changes take to number less than 0', async () => {
+    const [onChange] = await render({} as AssetVariableQuery);
+    const take = screen.getByRole('spinbutton');
+    onChange.mockClear();
+
+    await userEvent.clear(take);
+    await userEvent.tab();
+    await waitFor(() => {
+        expect(screen.getByText('Enter a value greater than or equal to 0')).toBeInTheDocument();
+        expect(onChange).toHaveBeenCalledWith(expect.objectContaining({ take: undefined }));
+    });
+})
+
+it('should not display error message when user changes value to number between 0 and max take', async () => {
+    const [onChange] = await render({} as AssetVariableQuery);
+    const take = screen.getByRole('spinbutton');
+    onChange.mockClear();
+
+    await userEvent.clear(take);
+    await userEvent.type(take, '5')
+    await userEvent.tab();
+    await waitFor(() => {
+        expect(screen.queryByText('Enter a value greater than or equal to 0')).not.toBeInTheDocument();
+        expect(screen.queryByText('Enter a value less than or equal to 10,000')).not.toBeInTheDocument();
+    });
+})
 
 it('renders the return type selector', async () => {
     render({  refId: '', type: AssetQueryType.ListAssets, filter: "" } as AssetQuery);