+++ src/datasources/asset/defaults.ts
@@ -1,3 +1,5 @@
+import { QUERY_LIMIT } from "./constants/constants"
+import { OutputType } from "./types/ListAssets.types"
 
 export const defaultAssetSummaryQuery = {
 }
@@ -8,5 +10,12 @@ export const defaultCalibrationForecastQuery = {
 }
 
 export const defaultListAssetsQuery = {
-    filter: ""
+    filter: "",
+    outputType: OutputType.Properties,
+    take: QUERY_LIMIT,
+}
+
+export const defaultListAssetsVariable = {
+    filter: "",
+    take: QUERY_LIMIT,
 }