+++ src/datasources/results/query-handlers/query-steps/QueryStepsDataSource.ts
@@ -20,6 +20,7 @@ import {
   StepsPropertiesOptions,
   StepsResponseProperties,
   InputOutputValues,
+  StepPath,
 } from 'datasources/results/types/QuerySteps.types';
 import { ResultsDataSourceBase } from 'datasources/results/ResultsDataSourceBase';
 import { defaultStepsQuery } from 'datasources/results/defaultQueries';
@@ -301,11 +302,16 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
     }
   }
 
-  async getStepPaths(resultsQuery: string): Promise<string[]> {
+  async getStepPaths(resultsQuery: string): Promise<StepPath[]> {
     if (resultsQuery) {
       const query = this.transformQuery(resultsQuery, this.resultsComputedDataFields, this.scopedVars);
       const stepPaths = await this.getStepPathsLookupValues(query!);
-      return this.flattenAndDeduplicate(stepPaths);
+      return this.flattenAndDeduplicate(stepPaths)
+      .filter(path => path && path.trim() !== '')
+      .map(path => ({
+        label: path?.replace(/\n/g, '\\') || '',
+        value: path || ''
+      }));
     }
     return [];
   }